Removed floating point stuff from kernel.
We can't use it there that easily because we would have to
save the user-state first and restore it afterwards. So, better
don't use it at all.
To ensure that we don't use it by accident in the future, I added
the -msoft-float option which results in a linker error if we
have floating point arithmetic.
For that reason, we had to move some stuff from basic into the
c library and some tests out of the kernel test into libctest.

+++ source/include/sys/mem/virtmem.h
@@ -215,9 +215,9 @@ public:
 	 * one usage of the region).
 	 *
 	 * @param pages will point to the number of pages (size of virtual-memory)
-	 * @return the number of used frames for this process
+	 * @return the used memory for this process in bytes
 	 */
-	float getMemUsage(size_t *pages) const;
+	size_t getMemUsage(size_t *pages) const;
 
 	/**
 	 * Requests the information we need to share the file that is mapped at <addr>. That is, the
